+++ src/lib.rs
@@ -125,7 +125,7 @@
     unsafe_op_in_unsafe_fn
 )]
 
-use std::{io::ErrorKind, mem::size_of};
+use std::mem::size_of;
 
 // Re-export the embedded_can crate so that applications can rely on
 // finding the same version we use.
@@ -204,68 +204,4 @@ pub(crate) fn as_bytes<T: Sized>(val: &T) -> &[u8] {
 pub(crate) fn as_bytes_mut<T: Sized>(val: &mut T) -> &mut [u8] {
     let sz = size_of::<T>();
     unsafe { std::slice::from_raw_parts_mut(val as *mut _ as *mut u8, sz) }
-}
-
-// ===== embedded_can I/O traits =====
-
-impl embedded_can::blocking::Can for CanSocket {
-    type Frame = CanFrame;
-    type Error = Error;
-
-    /// Blocking call to receive the next frame from the bus.
-    ///
-    /// This block and wait for the next frame to be received from the bus.
-    /// If an error frame is received, it will be converted to a `CanError`
-    /// and returned as an error.
-    fn receive(&mut self) -> Result<Self::Frame> {
-        use CanFrame::*;
-        match self.read_frame() {
-            Ok(Error(frame)) => Err(frame.into_error().into()),
-            Ok(frame) => Ok(frame),
-            Err(e) => Err(e.into()),
-        }
-    }
-
-    /// Blocking transmit of a frame to the bus.
-    fn transmit(&mut self, frame: &Self::Frame) -> Result<()> {
-        self.write_frame_insist(frame).map_err(|err| err.into())
-    }
-}
-
-impl embedded_can::nb::Can for CanSocket {
-    type Frame = CanFrame;
-    type Error = Error;
-
-    /// Non-blocking call to receive the next frame from the bus.
-    ///
-    /// If an error frame is received, it will be converted to a `CanError`
-    /// and returned as an error.
-    /// If no frame is available, it returns a `WouldBlck` error.
-    fn receive(&mut self) -> nb::Result<Self::Frame, Self::Error> {
-        use CanFrame::*;
-        match self.read_frame() {
-            Ok(Data(frame)) => Ok(Data(frame)),
-            Ok(Remote(frame)) => Ok(Remote(frame)),
-            Ok(Error(frame)) => Err(crate::Error::from(frame.into_error()).into()),
-            Err(err) => Err(match err.kind() {
-                ErrorKind::WouldBlock => nb::Error::WouldBlock,
-                _ => crate::Error::from(err).into(),
-            }),
-        }
-    }
-
-    /// Non-blocking transmit of a frame to the bus.
-    fn transmit(&mut self, frame: &Self::Frame) -> nb::Result<Option<Self::Frame>, Self::Error> {
-        match self.write_frame(frame) {
-            Ok(_) => Ok(None),
-            Err(err) => {
-                match err.kind() {
-                    ErrorKind::WouldBlock => Err(nb::Error::WouldBlock),
-                    // TODO: How to indicate buffer is full?
-                    // ErrorKind::StorageFull => Ok(frame),
-                    _ => Err(Error::from(err).into()),
-                }
-            }
-        }
-    }
 }